added unit tests for ConfirmationsThreshold
